development_20181101_DoanhNV: update question API

+++ NihongoServer/src/main/java/com/nihongo/service/data/impl/MCQQuestionServiceImpl.java
@@ -5,18 +5,28 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.nihongo.data.dao.ExamDAO;
 import com.nihongo.data.dao.MCQQuestionDAO;
 import com.nihongo.data.entity.AbstractEntity;
 import com.nihongo.data.entity.other.transfer.SearchData;
 import com.nihongo.data.entity.question.Question;
 import com.nihongo.dto.httpdto.request.AbstractSearchRequest;
+import com.nihongo.exception.AbstractNihongoException;
 import com.nihongo.service.data.MCQQuestionService;
+import com.nihongo.support.constant.ResponseCode;
 
+/**
+ * 
+ * @author DoanhNV Nov 1, 2018 - 10:29:03 PM
+ *
+ */
 @Service
 public class MCQQuestionServiceImpl implements MCQQuestionService {
 	
 	@Autowired
 	private MCQQuestionDAO mCQQuestionDAO;
+	@Autowired
+	private ExamDAO examDAO;
 	
 	@Override
 	public String insert(AbstractEntity entity) {
@@ -30,11 +40,15 @@ public class MCQQuestionServiceImpl implements MCQQuestionService {
 
 	@Override
 	public Question getById(String id) {
-		return null;
+		Question question = (Question) mCQQuestionDAO.getById(id);
+		return question;
 	}
 
 	@Override
 	public String delete(String id) {
+		if(examDAO.isExistsQuestion(id)) {
+			throw new AbstractNihongoException(ResponseCode.QUESTION_IS_IN_EXAM);
+		}
 		return mCQQuestionDAO.delete(id);
 	}
